@@ -2,12 +2,20 @@
   "name": "@dargmuesli/nuxt-cookie-control",
   "version": "2.0.0",
   "description": "Nuxt Cookies Control Module",
-  "license": "MIT",
   "author": "Dario Ferderber <dario.ferderber@broj42.com>",
   "contributors": [
     "Davor Teskera",
     "Jonas Thelemann"
   ],
+  "keywords": [
+    "nuxt cookie control",
+    "nuxt gdpr",
+    "nuxt cookie consent",
+    "nuxt gdpr consent"
+  ],
+  "repository": "https://github.com/dargmuesli/nuxt-cookie-control",
+  "license": "MIT",
+  "type": "module",
   "exports": {
     ".": {
       "import": "./dist/module.mjs",
@@ -16,14 +24,10 @@
   },
   "main": "./dist/module.cjs",
   "types": "./dist/module.d.ts",
-  "repository": "https://github.com/dargmuesli/nuxt-cookie-control",
-  "publishConfig": {
-    "access": "public"
-  },
   "files": [
     "dist",
-    "components",
-    "locale"
+    "src/components",
+    "src/locale"
   ],
   "scripts": {
     "prepack": "nuxt-module-build",
@@ -36,7 +40,7 @@
     "test": "yarn vitest run"
   },
   "dependencies": {
-    "@nuxt/kit": "3.0.0-rc.8",
+    "@nuxt/kit": "3.0.0-rc.9",
     "css-vars-ponyfill": "2.4.8",
     "sass": "1.54.6",
     "sass-loader": "13.0.2",
@@ -44,17 +48,19 @@
   },
   "devDependencies": {
     "@nuxt/module-builder": "0.1.7",
-    "@nuxt/test-utils": "npm:@nuxt/test-utils-edge@3.0.0-rc.8-27696168.75053c2",
-    "@nuxtjs/eslint-config-typescript": "10.0.0",
+    "@nuxtjs/eslint-config-typescript": "11.0.0",
     "@release-it/conventional-changelog": "5.1.0",
-    "c8": "7.12.0",
-    "consola": "2.15.3",
     "eslint": "8.23.0",
     "eslint-config-prettier": "8.5.0",
-    "nuxt": "3.0.0-rc.8",
-    "release-it": "15.4.0",
-    "typescript": "4.5.5",
-    "vue": "3.2.37",
+    "eslint-plugin-prettier": "4.2.1",
+    "husky": "8.0.1",
+    "lint-staged": "13.0.3",
+    "nuxt": "3.0.0-rc.9",
+    "prettier": "2.7.1",
+    "release-it": "15.4.1",
+    "typescript": "4.8.3",
+    "vite": "3.1.0",
+    "vue": "3.2.39",
     "webpack": "5.74.0"
   },
   "peerDependencies": {
@@ -63,5 +69,8 @@
   },
   "resolutions": {
     "@dargmuesli/nuxt-cookie-control": "link:./"
+  },
+  "publishConfig": {
+    "access": "public"
   }
 }
