@@ -1,5 +1,5 @@
 import { defineNuxtConfig } from 'nuxt'
 
 export default defineNuxtConfig({
-  modules: ['@dargmuesli/nuxt-cookie-control']
+  modules: ['@dargmuesli/nuxt-cookie-control'],
 })
